+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+**Added**
+
+- `Scribe.Encoder` protocol to derive a list of table headers from a struct.
+
+**Changed**
+
+- Bumped minimum Elixir version to 1.14.
+- More descriptive exception message when generating a table for a non-map/struct.
+- Better organization of documentation.
+
 ## v0.11.0 - 2024-08-31
 
 **Added**